+++ scripts/FCT/Linux/run_256ch_data_taking.sh
@@ -16,15 +16,17 @@ echo "DATADIR: "$DATADIR
 # Define the command to run the GUI script
 command="Sync.RunScriptArgs(\"/home/neutrino/FCT/code/scripts/FCT/Linux/Script_FCT_openshort.cs\",$sn,$bl1,$bl2)"
 command_citi="Sync.RunScriptArgs(\"/home/neutrino/FCT/code/scripts/FCT/Linux/Script_FCT_CITI_test.cs\",$sn)"
-# Close all GUIs to avoid double serial com
-sudo kill $(pidof mono)
-# Open GUI and wait 
+
+# Opens GUI only if there are no GUI already open
+if [ -z $(pidof mono) ]
+then 
 ( cd $GUI_path && mono $GUI_path$GUI_exe & )
 echo "Opening GUI ..."
 sleep 0.5
 echo "When GUI is open, press Enter "
 echo "(Close pop-up error windows on GUI, if any. DO NOT CLOSE THE SOCKET WINDOW! )"
 read -n 1
+fi
 
 # Open the serial com and send the command. Wait for it to end. Send second command, wait for it to end an close serial port com
 { sleep 1; echo $command; bash wait.sh $Data_path$dummy_EOS; sleep 1; echo $command_citi; bash wait.sh $Data_path$dummy_EOS_citi; } | telnet $ip_address $port 
@@ -32,5 +34,9 @@ read -n 1
 # sleep 1; echo $command; bash wait.sh $Data_path$dummy_EOS;
 
 # Close the GUI
-sudo kill $(pidof mono)
-sleep 1
+if [ -z $(pidof mono) ]
+then 
+    echo
+else
+    sudo kill $(pidof mono)
+fi
